Custom Car Physics Update In Vehicle Prefab
Got Suspension Working.
Fixed a Bug with using local coordinates in world space raycast that was causing the raycast to be offset different amounts as the car turned.
Fixed a bug that caused the spring strength to have an inverse effect where a thicker spring would compress easier. Now spring Strength is directly proportional to the distance it compresses.

Current Bug is the thin frame of the car which makes it top heavy causing it to tip over when turning.