Fixed a big bug with wheels.
The issue was in the Suspension script where I was raycasting from asymetical position on the left and right sides which caused the forces to apply unexpected torque values on the z axis.

One more bug with how the RPM works.

I need to calculate the RPM based on the current wheel Velocity.