Updated some Engine Code
Bug, The RPM measurement was independent of the wheel's angular velocity which didn't correct when external forces altered the car's speed. This meant the car's power potential didn't make sense for its current speed and yielded diminishing returns.

The solution was to calculate the wheel's angular velocity and base the theoretical Engine RPM based on that speed.

Solution is incomplete, I'm still working on the final details on how to convert RPM into Power at the wheels.