Fixed Bug with AI
When I increased the field of view to detect more cars in the blind spot, I also detected more Tracknodes which caused the average tracknode position to be extremely close to the AI car, which caused the AI to calculate extreme turn angles.

I Separated the field of view used to detect track nodes and other cars.

I also changed the code a bit to calculate offset positions rather than trying to offset the angle the car turns.